Machine/ Deep learning: Double-view CNN
Double-view CNN walkthrough: Instructions on how to run the program and what changes to be made.
dtyu.zip: Programs for generating synthetic x-ray scattering image, pre-processing the data for training and testing of the deep learning system, and training and testing the data. Instructions are in the Double-view CNN walkthrough pdf.